Did a full f*ing refactor
