+++ src/main/java/org/R4ZXRN3T/BottomToolBar.java
@@ -13,10 +13,10 @@ class BottomToolBar {
 
 	// only here to be able to refresh the button
 	private CustomButton undoButton;
-	private final Main main;
+	private final Firstpass firstpass;
 
-	public BottomToolBar(Main main) {
-		this.main = main;
+	public BottomToolBar(Firstpass firstpass) {
+		this.firstpass = firstpass;
 	}
 
 	// get and populate the toolbar
@@ -29,22 +29,22 @@ class BottomToolBar {
 		buttonPanel.setLayout(new FlowLayout(FlowLayout.LEFT));
 
 		// addButton
-		buttonPanel.add(new CustomButton("Add", main.isDarkMode() ? ADD_ICON_WHITE_SCALED : ADD_ICON_SCALED, e -> main.addAccount(), DEFAULT_BUTTON_SIZE));
+		buttonPanel.add(new CustomButton("Add", Config.getDarkMode() ? ADD_ICON_WHITE_SCALED : ADD_ICON_SCALED, _ -> firstpass.addAccount(), DEFAULT_BUTTON_SIZE));
 
 		// removeButton
-		buttonPanel.add(new CustomButton("Remove", main.isDarkMode() ? REMOVE_ICON_WHITE_SCALED : REMOVE_ICON_SCALED, e -> main.removeAccount(main.getTable().getSelectedRow()), main.getTable(), DEFAULT_BUTTON_SIZE));
+		buttonPanel.add(new CustomButton("Remove", Config.getDarkMode() ? REMOVE_ICON_WHITE_SCALED : REMOVE_ICON_SCALED, _ -> firstpass.removeAccount(firstpass.getTable().getSelectedRow()), firstpass.getTable(), DEFAULT_BUTTON_SIZE));
 
 		// editButton
-		buttonPanel.add(new CustomButton("Edit", main.isDarkMode() ? EDIT_ICON_WHITE_SCALED : EDIT_ICON_SCALED, e -> main.editAccount(main.getTable().getSelectedRow()), main.getTable(), DEFAULT_BUTTON_SIZE));
+		buttonPanel.add(new CustomButton("Edit", Config.getDarkMode() ? EDIT_ICON_WHITE_SCALED : EDIT_ICON_SCALED, _ -> firstpass.editAccount(firstpass.getTable().getSelectedRow()), firstpass.getTable(), DEFAULT_BUTTON_SIZE));
 		// undoButton
-		undoButton = new CustomButton("Undo", main.isDarkMode() ? UNDO_ICON_WHITE_SCALED : UNDO_ICON_SCALED, e -> main.undoDeletion(), DEFAULT_BUTTON_SIZE);
+		undoButton = new CustomButton("Undo", Config.getDarkMode() ? UNDO_ICON_WHITE_SCALED : UNDO_ICON_SCALED, _ -> firstpass.undoDeletion(), DEFAULT_BUTTON_SIZE);
 		buttonPanel.add(undoButton);
 		refreshUndoButton();
 		// settingsButton
-		buttonPanel.add(new CustomButton("Generator", main.isDarkMode() ? GENERATE_ICON_WHITE_SCALED : GENERATE_ICON_SCALED, e -> PasswordGenerator.showPasswordGeneratorDialog(main.getFrame(), main.isDarkMode()), DEFAULT_BUTTON_SIZE));
+		buttonPanel.add(new CustomButton("Generator", Config.getDarkMode() ? GENERATE_ICON_WHITE_SCALED : GENERATE_ICON_SCALED, _ -> PasswordGenerator.showPasswordGeneratorDialog(firstpass.getFrame(), Config.getDarkMode()), DEFAULT_BUTTON_SIZE));
 		// exitButton
 		JPanel exitPanel = new JPanel(new FlowLayout(FlowLayout.RIGHT));
-		exitPanel.add(new CustomButton("Save & Exit", main.isDarkMode() ? EXIT_ICON_WHITE_SCALED : EXIT_ICON_SCALED, e -> main.exit(), new Dimension(125, 35)), BorderLayout.EAST);
+		exitPanel.add(new CustomButton("Save & Exit", Config.getDarkMode() ? EXIT_ICON_WHITE_SCALED : EXIT_ICON_SCALED, _ -> firstpass.exit(), new Dimension(125, 35)), BorderLayout.EAST);
 
 		toolBar.add(buttonPanel, BorderLayout.WEST);
 		toolBar.add(exitPanel, BorderLayout.EAST);
@@ -55,6 +55,6 @@ class BottomToolBar {
 	// refresh the undo button
 	// this damn button is annoying
 	public void refreshUndoButton() {
-		undoButton.setEnabled(!main.getUndoStack().isEmpty());
+		undoButton.setEnabled(!firstpass.getUndoStack().isEmpty());
 	}
 }
