+++ src/main/java/org/R4ZXRN3T/SettingsMenu.java
@@ -1,7 +1,5 @@
 package org.R4ZXRN3T;
 
-import org.R4ZXRN3T.Files.ConfigKey;
-
 import javax.swing.*;
 import java.awt.*;
 import java.util.HashMap;
@@ -13,10 +11,10 @@ class SettingsMenu {
 	private boolean needsRestart = false;
 	private CustomButton changePasswordButton;
 	private CustomButton removePasswordButton;
-	private final Main main;
+	private final Firstpass firstpass;
 
-	public SettingsMenu(Main main) {
-		this.main = main;
+	public SettingsMenu(Firstpass firstpass) {
+		this.firstpass = firstpass;
 	}
 
 	public void showSettings() {
@@ -24,7 +22,7 @@ class SettingsMenu {
 		setCurrentSettings();
 
 		// set up frame
-		settingsFrame = new JDialog(main.getFrame(), "Firstpass Settings", true);
+		settingsFrame = new JDialog(firstpass.getFrame(), "Firstpass Settings", true);
 		settingsFrame.setLayout(new BorderLayout(16, 16));
 		settingsFrame.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
 
@@ -35,7 +33,7 @@ class SettingsMenu {
 		// set up the frame
 		settingsFrame.setSize(500, 400);
 		settingsFrame.setResizable(false);
-		settingsFrame.setLocationRelativeTo(main.getFrame());
+		settingsFrame.setLocationRelativeTo(firstpass.getFrame());
 		settingsFrame.setIconImage(Icons.FIRSTPASS_ICON.getImage());
 		settingsFrame.requestFocus();
 		settingsFrame.setVisible(true);
@@ -43,9 +41,9 @@ class SettingsMenu {
 
 	// write the current settings to the HashMap
 	private void setCurrentSettings() {
-		currentSettings.put(0, main.getCorrectPassword());
-		currentSettings.put(1, Files.getConfig(ConfigKey.LOOK_AND_FEEL));
-		currentSettings.put(2, Files.getConfig(ConfigKey.CHECK_FOR_UPDATES));
+		currentSettings.put(0, firstpass.getCorrectPassword());
+		currentSettings.put(1, Config.getConfig(Config.ConfigKey.LOOK_AND_FEEL));
+		currentSettings.put(2, Config.getConfig(Config.ConfigKey.CHECK_FOR_UPDATES));
 	}
 
 	// get the panel for theme settings
@@ -56,7 +54,7 @@ class SettingsMenu {
 		String[] themeOptions = {"Flat Light", "Flat Dark", "Flat Mac Light", "Flat Mac Dark", "Flat IntelliJ", "Flat Darcula", "Swing Metal", "System Default"};
 		JComboBox<String> themeSelector = new JComboBox<>(themeOptions);
 		themeSelector.setSelectedIndex(Integer.parseInt(currentSettings.get(1)));
-		themeSelector.addActionListener(e -> {
+		themeSelector.addActionListener(_ -> {
 			currentSettings.replace(1, String.valueOf(themeSelector.getSelectedIndex()));
 			needsRestart = true;
 		});
@@ -69,12 +67,12 @@ class SettingsMenu {
 	private JPanel getPasswordPanel() {
 		JPanel passwordPanel = new JPanel(new FlowLayout(FlowLayout.LEFT));
 		passwordPanel.setBorder(BorderFactory.createTitledBorder("Password"));
-		changePasswordButton = new CustomButton("Change Password", e -> {
+		changePasswordButton = new CustomButton("Change Password", _ -> {
 			changePassword();
 			refreshButton();
 		}, new Dimension(140, 30));
 		passwordPanel.add(changePasswordButton);
-		removePasswordButton = new CustomButton("Remove Password", e -> removePasswordDialog(), new Dimension(140, 30));
+		removePasswordButton = new CustomButton("Remove Password", _ -> removePasswordDialog(), new Dimension(140, 30));
 		passwordPanel.add(removePasswordButton);
 		refreshButton();
 		return passwordPanel;
@@ -83,7 +81,7 @@ class SettingsMenu {
 	private JPanel getFullDeletePane() {
 		JPanel deletePanel = new JPanel(new FlowLayout(FlowLayout.LEFT));
 		deletePanel.setBorder(BorderFactory.createTitledBorder("Delete"));
-		deletePanel.add(new CustomButton("Delete Everything", null, e -> fullDeleteDialog(), false, true, new Dimension(140, 30), null, Color.RED, null));
+		deletePanel.add(new CustomButton("Delete Everything", null, _ -> fullDeleteDialog(), false, true, new Dimension(140, 30), null, Color.RED, null));
 		JLabel deleteLabel = new JLabel("Delete all data");
 		deletePanel.add(deleteLabel);
 		return deletePanel;
@@ -93,10 +91,10 @@ class SettingsMenu {
 		JPanel updatePanel = new JPanel(new FlowLayout(FlowLayout.LEFT));
 		updatePanel.setBorder(BorderFactory.createTitledBorder("Update"));
 
-		updatePanel.add(new CustomButton("Check for Updates", e -> {
-			main.setUpdateAvailable(Updater.checkVersion(true).compareToIgnoreCase(Main.CURRENT_VERSION) > 0);
-			main.getTopToolBar().getUpdateButton().setVisible(main.isUpdateAvailable());
-			if (main.isUpdateAvailable()) {
+		updatePanel.add(new CustomButton("Check for Updates", _ -> {
+			firstpass.setUpdateAvailable(Updater.checkVersion(true).compareToIgnoreCase(Firstpass.CURRENT_VERSION) > 0);
+			firstpass.getTopToolBar().getUpdateButton().setVisible(firstpass.isUpdateAvailable());
+			if (firstpass.isUpdateAvailable()) {
 				Updater.update();
 			} else {
 				JOptionPane.showMessageDialog(settingsFrame, "You are already using the latest version.", "No Updates Available", JOptionPane.INFORMATION_MESSAGE);
@@ -105,9 +103,7 @@ class SettingsMenu {
 
 		JCheckBox updateCheckBox = new JCheckBox("Check for updates on startup");
 		updateCheckBox.setSelected(Boolean.parseBoolean(currentSettings.get(2)));
-		updateCheckBox.addActionListener(e -> {
-			currentSettings.replace(2, String.valueOf(updateCheckBox.isSelected()));
-		});
+		updateCheckBox.addActionListener(_ -> currentSettings.replace(2, String.valueOf(updateCheckBox.isSelected())));
 		updatePanel.add(updateCheckBox);
 
 		return updatePanel;
@@ -118,11 +114,11 @@ class SettingsMenu {
 		JToolBar bottomToolbar = new JToolBar();
 		bottomToolbar.setLayout(new FlowLayout(FlowLayout.RIGHT));
 		bottomToolbar.setBorder(BorderFactory.createEmptyBorder(16, 16, 16, 16));
-		bottomToolbar.add(new CustomButton("Apply", main.isDarkMode() ? Icons.APPLY_ICON_WHITE_SCALED : Icons.APPLY_ICON_SCALED, e -> applySettings(), false, false, new Dimension(90, 30), main.isDarkMode() ? new Color(50, 50, 50) : Color.lightGray, null, null));
-		bottomToolbar.add(new CustomButton("Cancel", main.isDarkMode() ? Icons.CANCEL_ICON_WHITE_SCALED : Icons.CANCEL_ICON_SCALED, e -> {
-			main.getFrame().setEnabled(true);
+		bottomToolbar.add(new CustomButton("Apply", Config.getDarkMode() ? Icons.APPLY_ICON_WHITE_SCALED : Icons.APPLY_ICON_SCALED, _ -> applySettings(), false, false, new Dimension(90, 30), Config.getDarkMode() ? new Color(50, 50, 50) : Color.lightGray, null, null));
+		bottomToolbar.add(new CustomButton("Cancel", Config.getDarkMode() ? Icons.CANCEL_ICON_WHITE_SCALED : Icons.CANCEL_ICON_SCALED, _ -> {
+			firstpass.getFrame().setEnabled(true);
 			settingsFrame.dispose();
-		}, false, false, new Dimension(90, 30), main.isDarkMode() ? new Color(50, 50, 50) : Color.lightGray, null, null));
+		}, false, false, new Dimension(90, 30), Config.getDarkMode() ? new Color(50, 50, 50) : Color.lightGray, null, null));
 		return bottomToolbar;
 	}
 
@@ -144,17 +140,17 @@ class SettingsMenu {
 		JTextField oldPassword = new JTextField();
 		JTextField newPassword = new JTextField();
 
-		Object[] message = main.isPasswordSet() ? new Object[]{"Old Password:", oldPassword, "New Password:", newPassword} : new Object[]{"New Password:", newPassword};
+		Object[] message = firstpass.isPasswordSet() ? new Object[]{"Old Password:", oldPassword, "New Password:", newPassword} : new Object[]{"New Password:", newPassword};
 
 		int option = JOptionPane.showConfirmDialog(settingsFrame, message, "Change Password", JOptionPane.OK_CANCEL_OPTION);
 		if (option == JOptionPane.OK_OPTION) {
-			if (!main.isPasswordSet()) {
+			if (!firstpass.isPasswordSet()) {
 				currentSettings.replace(0, newPassword.getText());
-				main.setPasswordSet(true);
+				firstpass.setPasswordSet(true);
 				JOptionPane.showMessageDialog(settingsFrame, "Password successfully set", "Success", JOptionPane.INFORMATION_MESSAGE);
 				return;
 			}
-			if (oldPassword.getText().equals(main.getCorrectPassword())) {
+			if (oldPassword.getText().equals(firstpass.getCorrectPassword())) {
 				currentSettings.replace(0, newPassword.getText());
 				JOptionPane.showMessageDialog(settingsFrame, "Password successfully changed", "Success", JOptionPane.INFORMATION_MESSAGE);
 			} else {
@@ -165,21 +161,21 @@ class SettingsMenu {
 
 	// apply the settings and restart the program if necessary
 	private void applySettings() {
-		main.setCorrectPassword(currentSettings.get(0));
-		Files.setConfig(ConfigKey.LOOK_AND_FEEL, currentSettings.get(1));
-		Files.setConfig(ConfigKey.CHECK_FOR_UPDATES, currentSettings.get(2));
+		firstpass.setCorrectPassword(currentSettings.get(0));
+		Config.setConfig(Config.ConfigKey.LOOK_AND_FEEL, currentSettings.get(1));
+		Config.setConfig(Config.ConfigKey.CHECK_FOR_UPDATES, currentSettings.get(2));
 
 		if (needsRestart) {
 			String message = "The program needs to be restarted in order to apply the new settings. Do you want to restart now?";
 			int option = JOptionPane.showConfirmDialog(null, message, "Restart to apply settings", JOptionPane.YES_NO_OPTION);
 			if (option == JOptionPane.YES_OPTION) {
-				main.save();
-				Tools.restart();
+				firstpass.save();
+				Main.restart(firstpass);
 			}
 		}
 
 		settingsFrame.dispose();
-		main.setChangeMade(true);
+		firstpass.setChangeMade(true);
 	}
 
 	private void fullDeleteDialog() {
@@ -188,7 +184,7 @@ class SettingsMenu {
 		warningLabel2.setForeground(Color.RED);
 		int option = JOptionPane.showConfirmDialog(null, new Object[]{warningLabel, warningLabel2}, "Delete all data", JOptionPane.YES_NO_OPTION);
 		if (option == JOptionPane.YES_OPTION) {
-			main.fullDelete();
+			firstpass.fullDelete();
 		}
 	}
 
@@ -198,13 +194,13 @@ class SettingsMenu {
 			return;
 		}
 		currentSettings.replace(0, "");
-		main.setPasswordSet(false);
+		firstpass.setPasswordSet(false);
 		JOptionPane.showMessageDialog(null, "Password removed", "Success", JOptionPane.INFORMATION_MESSAGE);
 		refreshButton();
 	}
 
 	private void refreshButton() {
-		changePasswordButton.setText(main.isPasswordSet() ? "Change Password" : "Set Password");
-		removePasswordButton.setEnabled(main.isPasswordSet());
+		changePasswordButton.setText(firstpass.isPasswordSet() ? "Change Password" : "Set Password");
+		removePasswordButton.setEnabled(firstpass.isPasswordSet());
 	}
 }
