+++ src/main/java/org/R4ZXRN3T/TopToolBar.java
@@ -3,42 +3,42 @@ package org.R4ZXRN3T;
 import javax.swing.*;
 import java.awt.*;
 
-class TopToolBar {
+public class TopToolBar {
 
-	private JButton updateButton = new JButton("New version available!");
-	private final Main main;
+	private final JButton updateButton = new JButton("New version available!");
+	private final Firstpass firstpass;
 
-	public TopToolBar(Main main) {
-		this.main = main;
+	public TopToolBar(Firstpass firstpass) {
+		this.firstpass = firstpass;
 	}
 
 	public JMenuBar getTopToolBar() {
 		JMenuBar toolBar = new JMenuBar();
 
 		JMenuItem saveItem = new JMenuItem("Save");
-		saveItem.addActionListener(e -> {
-			main.save();
-			main.setChangeMade(false);
+		saveItem.addActionListener(_ -> {
+			firstpass.save();
+			firstpass.setChangeMade(false);
 		});
-		saveItem.setIcon(main.isDarkMode() ? Icons.SAVE_ICON_WHITE_SCALED : Icons.SAVE_ICON_SCALED);
+		saveItem.setIcon(Config.getDarkMode() ? Icons.SAVE_ICON_WHITE_SCALED : Icons.SAVE_ICON_SCALED);
 
 		JMenuItem exportAsItem = new JMenuItem("Export");
-		exportAsItem.addActionListener(e -> Files.exportData(main));
-		exportAsItem.setIcon(main.isDarkMode() ? Icons.EXPORT_ICON_WHITE_SCALED : Icons.EXPORT_ICON_SCALED);
+		exportAsItem.addActionListener(_ -> Files.exportData(firstpass));
+		exportAsItem.setIcon(Config.getDarkMode() ? Icons.EXPORT_ICON_WHITE_SCALED : Icons.EXPORT_ICON_SCALED);
 
 		JMenuItem importAsItem = new JMenuItem("Import");
-		importAsItem.setIcon(main.isDarkMode() ? Icons.IMPORT_ICON_WHITE_SCALED : Icons.IMPORT_ICON_SCALED);
-		importAsItem.addActionListener(e -> Files.importData(main));
+		importAsItem.setIcon(Config.getDarkMode() ? Icons.IMPORT_ICON_WHITE_SCALED : Icons.IMPORT_ICON_SCALED);
+		importAsItem.addActionListener(_ -> Files.importData(firstpass));
 
 		JButton settingsButton = new JButton("Settings");
-		settingsButton.addActionListener(e -> new SettingsMenu(main).showSettings());
-		settingsButton.setBackground(main.getFrame().getBackground());
+		settingsButton.addActionListener(_ -> new SettingsMenu(firstpass).showSettings());
+		settingsButton.setBackground(firstpass.getFrame().getBackground());
 		settingsButton.setBorderPainted(false);
 		settingsButton.setFocusable(false);
 
 		JMenuItem exitItem = new JMenuItem("Save & Exit");
-		exitItem.addActionListener(e -> main.exit());
-		exitItem.setIcon(main.isDarkMode() ? Icons.EXIT_ICON_WHITE_SCALED : Icons.EXIT_ICON_SCALED);
+		exitItem.addActionListener(_ -> firstpass.exit());
+		exitItem.setIcon(Config.getDarkMode() ? Icons.EXIT_ICON_WHITE_SCALED : Icons.EXIT_ICON_SCALED);
 
 		JMenu fileItem = new JMenu("File");
 		fileItem.add(saveItem);
@@ -48,8 +48,8 @@ class TopToolBar {
 
 		updateButton.setForeground(new Color(0, 180, 255));
 		updateButton.setBorderPainted(true);
-		updateButton.addActionListener(e -> Updater.update());
-		updateButton.setVisible(main.isUpdateAvailable());
+		updateButton.addActionListener(_ -> Updater.update());
+		updateButton.setVisible(firstpass.isUpdateAvailable());
 
 		toolBar.add(fileItem);
 		toolBar.add(settingsButton);
